+++ soc/silabs/silabs_s2/soc_radio.h
@@ -1,0 +1,13 @@
+/*
+ * Copyright (c) 2025 Silicon Laboratories Inc.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+/**
+ * @file
+ * @brief Declare radio functions for Silicon Labs Series 2 products
+ *
+ */
+
+void rail_isr_installer(void);
